fix: move ListNode to root of grok package
