+++ src/main/java/com/pursuetao/algs/grok/fastslow/S0.java
@@ -1,5 +1,7 @@
 package com.pursuetao.algs.grok.fastslow;
 
+import com.pursuetao.algs.grok.ListNode;
+
 // LinkedList Cycle
 // Given the head of a Singly LinkedList, write a function to determine if the LinkedList has a cycle in it or not.
 public class S0 {
