+++ src/main/java/com/pursuetao/algs/grok/fastslow/S2.java
@@ -1,5 +1,7 @@
 package com.pursuetao.algs.grok.fastslow;
 
+import com.pursuetao.algs.grok.ListNode;
+
 // Start of LinkedList Cycle
 // Given the head of a Singly LinkedList that contains a cycle, write a function to find the starting node of the cycle.
 public class S2 {
