+++ src/main/java/com/pursuetao/algs/grok/fastslow/S4.java
@@ -1,5 +1,7 @@
 package com.pursuetao.algs.grok.fastslow;
 
+import com.pursuetao.algs.grok.ListNode;
+
 // Middle of the LinkedList
 // Given the head of a Singly LinkedList, write a method to return the middle node of the LinkedList.
 // If the total number of nodes in the LinkedList is even, return the second middle node.
