+++ src/test/java/com/pursuetao/algs/grok/fastslow/TestS0.java
@@ -1,5 +1,6 @@
 package com.pursuetao.algs.grok.fastslow;
 
+import com.pursuetao.algs.grok.ListNode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
